Updates to allow code to run in newer versions of pytorch

+++ helper.py
@@ -208,7 +208,7 @@ class Helper:
             if self.params['diff_privacy']:
                 update_per_layer.add_(self.dp_noise(data, self.params['sigma']))
 
-            data.add_(update_per_layer)
+            data.add_(update_per_layer.long())
 
         return True
